Moved all errors to errors.py Moved add_reporter() to operations.py
