+++ python/infestor/errors.py
@@ -1,0 +1,6 @@
+class GenerateReporterError(Exception):
+    pass
+
+
+class GenerateDecoratorError(Exception):
+    pass
