+++ python/infestor/manager.py
@@ -1,42 +1,15 @@
 from typing import Tuple, List, Optional
 import os
 import libcst as cst
-import logging
 from . import visitors
 from . import transformers
-
+from .errors import *
 from .config import (
     default_config_file,
     load_config,
     save_config,
     python_root_relative_to_repository_root,
 )
-
-DEFAULT_REPORTER_FILENAME = "report.py"
-REPORTER_FILE_TEMPLATE: Optional[str] = None
-TEMPLATE_FILEPATH = os.path.join(os.path.dirname(__file__), "report.py.template")
-
-try:
-    with open(TEMPLATE_FILEPATH, "r") as ifp:
-        REPORTER_FILE_TEMPLATE = ifp.read()
-except Exception as e:
-    logging.warn(f"WARNING: Could not load reporter template from {TEMPLATE_FILEPATH}:")
-    logging.warn(e)
-
-# TODO(zomglings): Use an Enum here.
-CALL_TYPE_SYSTEM_REPORT = "system_report"
-CALL_TYPE_SETUP_EXCEPTHOOK = "setup_excepthook"
-
-DECORATOR_TYPE_RECORD_CALL = "record_call"
-DECORATOR_TYPE_RECORD_ERRORS = "record_errors"
-
-
-class GenerateReporterError(Exception):
-    pass
-
-
-class GenerateConfigurationError(Exception):
-    pass
 
 
 def get_reporter_module_path(
@@ -183,47 +156,3 @@ class PackageFileManager:
         modified_tree = self.syntax_tree.visit(transformer)
         self._visit(modified_tree)
 
-
-def add_reporter(
-    repository: str,
-    reporter_filepath: Optional[str] = None,
-    force: bool = False,
-) -> None:
-    if REPORTER_FILE_TEMPLATE is None:
-        raise GenerateReporterError("Could not load reporter template file")
-
-    config_file = default_config_file(repository)
-    configuration = load_config(config_file)
-
-    if reporter_filepath is None:
-        if configuration.reporter_filepath is not None:
-            reporter_filepath = configuration.reporter_filepath
-        else:
-            reporter_filepath = DEFAULT_REPORTER_FILENAME
-    else:
-        if (
-            configuration.reporter_filepath is not None
-            and configuration.reporter_filepath != reporter_filepath
-        ):
-            raise GenerateReporterError(
-                f"Configuration expects reporter to be set up at a different file than the one specified; specified={reporter_filepath}, expected={configuration.reporter_filepath}"
-            )
-
-    reporter_filepath_full = os.path.join(repository, reporter_filepath)
-    if (not force) and os.path.exists(reporter_filepath_full):
-        raise GenerateReporterError(
-            f"Object already exists at desired reporter filepath: {reporter_filepath_full}"
-        )
-
-    if configuration.reporter_token is None:
-        raise GenerateReporterError("No reporter token was specified in configuration")
-
-    contents = REPORTER_FILE_TEMPLATE.format(
-        project_name=configuration.project_name,
-        reporter_token=configuration.reporter_token,
-    )
-    with open(reporter_filepath_full, "w") as ofp:
-        ofp.write(contents)
-
-    configuration.reporter_filepath = reporter_filepath
-    save_config(config_file, configuration)
