+++ python/infestor/models.py
@@ -1,19 +1,19 @@
 from pydantic import BaseModel
 from typing import List
 
-# class.func.innerfn
+
 class ReporterCall(BaseModel):
     call_type: str
     lineno: int
-    scope_stack: List[str]
+    scope_stack: str
 
 
 class ReporterDecorator(BaseModel):
     decorator_type: str
     lineno: int
-    scope_stack: List[str]
+    scope_stack: str
 
 
 class ReporterDecoratorCandidate(BaseModel):
-    scope_stack: List[str]
+    scope_stack: str
     lineno: int
